Add support for Azure/sonic-buildimage-msft in PR Testing. (#17050)
What is the motivation for this PR?
Previously, the Download Pr Script step in PR testing did not support the Azure/sonic-buildimage-msft repository. This PR resolves the issue by adding the repository name to the if condition.

How did you do it?
This PR resolves the issue by adding the repository name to the if condition.

How did you verify/test it?

Change-Id: I2ae43908724c74f6e5648f9712f338eb7ba26006