Support downloading buildimage from 202412 branch.  (#17090)
What is the motivation for this PR?
The previous code retrieved all images from the Azure.sonic-buildimage pipeline, which does not include the 202412 branch. This PR resolves the issue by adding a conditional check. If the branch is 202412, the image will be downloaded from the Azure.sonic-buildimage-msft.PR pipeline (definitionId: 2511).

How did you do it?
This PR resolves the issue by adding a conditional check. If the branch is 202412, the image will be downloaded from the Azure.sonic-buildimage-msft.PR pipeline (definitionId: 2511).

How did you verify/test it?
Manually check the new added url.

Change-Id: Iab0170c17ae8f2ff9475b1712ac1f3ae4d12beb7