Change assertion to a `wait_until` in `test_route_withdraw_advertise` (#16858)
We were seeing that we'd get occasional AssertionErrors in test_route_withdraw_advertise when the routes were advertised because we didn't wait long enough before the assertion.

Instead of increasing the sleep time further I converted the assertion to a wait_until so we will only wait as long as we need to.

Note: When casting to 202405 we'll need to add the include for wait_until, it's not in this PR because it's already included in master from tests.common.utilities import wait_until
Change-Id: I38f60c1f727733a3dec7b11e711c9f16ea8e0c73