[test gap] Add a new case to check caclmgrd syslog (#17143)
What is the motivation for this PR?
A new test case to check if INFO logs of caclmgrd was printed into syslog successfully.

How did you do it?
rotate syslog
restart caclmgrd
check if iptables syslog existing in syslog file
Check if iptables rules are applied successfully in the output of systemctl status caclmgrd
How did you verify/test it?
Run the new test case on the testbed

Change-Id: Iae532b23a41b3bc85d49e7e3dcabac6129876584