Fix multiasic testgap for iface_namingmode (#17045)
Description of PR
Summary:

Add support for multi-asic compatible for this tests. This PR requires sonic-net/sonic-utilities#3781 to be merged as it's support -n <namespace>

Fixes # (issue) 28837145

Approach
What is the motivation for this PR?
Described above

How did you do it?
Utility improvement: sonic-net/sonic-utilities#3781

Use namespace to get the information so that it's compatible with multi-asic.

Enabled internal port when querying for port_alias_fact which was disabled before for compatibility

How did you verify/test it?
T2 testbed verified

Change-Id: I37781e52f5bae240c970a96c89644073df80f53c
Signed-off-by: Austin Pham <austinpham@microsoft.com>