Add a testcase for GNOI OS.Verify. (#16770)
Description of PR
Summary:
Add a testcase for GNOI OS.Verify.

Approach
What is the motivation for this PR?
Add a testcase for GNOI OS.Verify.

How did you do it?
Add a testcase for GNOI OS.Verify.

How did you verify/test it?
On physical switch:
with an up to date image

Change-Id: I53a5c9ced42e13df2f750c75fda9ba3546a60bd7