Update gnoi and mvrf test cases for chrony support (#17434)
What is the motivation for this PR?
With the NTP daemon being changed from ntpd to chrony, commands that refer to the ntpd daemon need to be updated to point to chrony instead.

How did you do it?
Update the gnoi and mvrf test cases to point to chrony-related commands instead of ntpd-related commands.

How did you verify/test it?
Tested on physical T0 and T2 testbed.

Change-Id: Ie0fa5c9570a0225c9559c98cd82fb5a37fab4f52
Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>