Add traffic tests for static and directly connected routes in IPFWD test suite (#17058)
What is the motivation for this PR?
Additional coverage in forwarding test

How did you do it?
Added extra cases to the existing script to handle the cases where traffic is sent to directly connected route as well as to a prefix that is reachable through a static route.

How did you verify/test it?
Run on T1 topology running master

Change-Id: I4055c622e5f835419ce384bc24103026b3e510a8