Adding static route and ECN changes needed to pass. (#17012)
Description of PR
Summary:
For cisco-8000, if the traffic flows to a connected route, the system uses the default voq. If we use static route, it uses the split-voq. In this PR, we change the dest IP address of the traffic streams to a static routed destination, and add that static route as well in the DUT. To test it, we have used the ECN script, and had to modify it as well.

In ECN we keep hitting the drop or other limits before hitting the xoff limits, due to the default voq. So changing it to static route helps in passing the tests.
Fixes # (issue)

Approach
What is the motivation for this PR?
Hitting failures in multiple tests due to traffic taking default voq (due to the connected routes).

How did you do it?
Added a new function to add static route in DUT, and use the new IP destination value for the traffic stream.

How did you verify/test it?
Alpesh had run the code in 202405 in my TB.

Any platform specific information?
Only for cisco-8000.

co-authorized by: jianquanye@microsoft.com

Change-Id: Icbe7f8dc29ab02215305b8439ce2c8b726d54774