Fail the test when evaluate conditional mark fails (#17522)
What is the motivation for this PR?
Previously, when condition evaluation failed due to a syntax error, we only logged an error message, making it difficult to identify such regressions. In this PR, we improve the code by ensuring that any syntax error in the condition will cause the test to fail, providing a clear and actionable error message.

How did you do it?
In this PR, we improve the code by ensuring that any syntax error in the condition will cause the test to fail, providing a clear and actionable error message.

How did you verify/test it?
Manually create a syntax error, and it will fail the whole test.

Change-Id: Icc1b0837e2f21dae2208847e562e808062532903