[performance_meter] allow specific run number for each test (#17609)
What is the motivation for this PR?
Originally run count of all performance meter test cases are the same and is input by command line argument from run_tests. It can be inconvenient sometimes

How did you do it?
Add entry in config file and read it. The config file will be assessed in pytest_generate_tests

Change-Id: If6e3aa6cdf863f32ea5b3a124dd4cd1ab5c7eaed