[pre-commit] Use correct comment format # noqa: not # noqa in tests/b* (#17759)
What is the motivation for this PR?
In pre-commit, we were using # noqa to ignore failures/errors/warnings in our code, but once we use # noqa, it will ignore all pre-commit errors in this line, the correct solution is using # noqa: so only specific failures/errors/warnings would be ignored.

How did you do it?
Replace # noqa with # noqa: in tests/b*

Change-Id: I668578a9e8b9254d0b6aa498e2d7acc59a670e6c