Remove the testbed.csv file (#4207872)
What is the motivation for this PR?
Remove the obsolete testbed.csv replaced by testbed.yaml. Community removed the file https://github.com/sonic-net/sonic-mgmt/pull/15064

How did you do it?
Deleted the file and the symlink to it

How did you verify/test it?
Confirmed that the file is not used with Cumulus and NVOS, testes the custom tarball w/o the file on the canonical and community test ted. See the redmine ticket #4207872

Change-Id: If0dc00b9e45f3f835aaed5172130bbc5fb4c2ae1