Fix loganalyzer regex to ignore chronyd related ERR logs (#17880)
What is the motivation for this PR?
PR #16239 ignored an error message coming from ntpd (actually written by nss_tacplus). As that PR explains, this can happen when a config reload is run, and chrony calls getpwnam when /etc/tacplus_nss.conf is still being regenerated. However, it is not a failure.

How did you do it?
With the change of NTP daemons, this loganalyzer ignore regex needs to be updated to match the right chrony log.

How did you verify/test it?
By running sonic-mgmt tests that perform config reload and verifying that these errors do not cause loganalyzer to mark the tests as failed.

Change-Id: I7e361b032f25b2a23ad20a822fa2c3816f39f198
Signed-off-by: Prabhat Aravind <paravind@microsoft.com>