Changes for VOQ single-dut multi-asic (#17852)
What is the motivation for this PR?
New testbed modification of VOQ Single DUT and multi-asic.
How did you do it?
If switch type is VOQ, decide supervisor based on the number of Duts and also based on if the DUT is modular chassis or not.
How did you verify/test it?

Change-Id: Ib3a060aaba3b893ae6ef0ec01a2e4787523c489e