Report the hwsku, platform, asic_type in tested facts from the switch
What is the motivation for this PR?
hwsku may be set to the different value on the switch during the
deployment using --dest-hwsku option.

How did you do it?
Read the value from the switch and report instead of the value
from the platform_params.
Also reporting platform, asic_type from the switch

How did you test it?
Ran the custom tarball on the mgmt container.

Change-Id: Ic7fa78697dbb49dab2b66bd9b7c3806bc007ef37