Add os port name to Mongo DB
Details:
adding the port name in the os to the mongo db, so it can be used
instead of the sdk port in the performance dashboard.
i.e, 'Ethernet0' instead of '0x100f1'.
remove tc data key from db in case of empty TC data(800G tests don't
support hft and don't collect this data), it's preferable not to have
keys with empty values.

Change-Id: Ib502868adf28366f33df22459262adabe0e4c69e