Repave testing infra for virtual chassis (#18064)
Description of PR
Summary: there were some issues with the infrastructure found during testing with virtual chassis, as articulated below:

BMP feature is enabled on latest master image with multi-asic LCs. However, it is not marked as DEFAULT_ASIC_SERVICES so we need to manually check it when doing modify_syslog_rate_limit.
The virtual chassis supervisor is not acting like a fabric card and it does not have fabric card's info so we should not mark its switch_type as fabric. I added a new switch_type called 'dummy-sup'.
Fixes # (issue)

Change-Id: Ib681b6f0d639876c7b9501e4fb0889f62fc9c755