Issue #17885: Update gNMI ConfigDB test to expect replace instead of reload (#17886)
Summary:
Fixes # 17885

Add a test case to gnmi/test_gnmi_configdb.py to test sending a full configuration file using SET-REPLACE in order to trigger a call to GCU's 'config replace' on the device.

This tests the codepath introduced here.

Approach
What is the motivation for this PR?
To support the changes to sonic-gnmi proposed here

How did you do it?
Add a new test case following the existing testcase that tests a full config reload (test_gnmi_configdb_full_01).

How did you verify/test it?
With the changes made to sonic-gnmi to use ConfigReplace for a full config update, the current test fails. This version passes with those changes.

Any platform specific information?
Supported testbed topology if it's a new test case?
Supports all testbeds supported by the test as a whole.

Change-Id: I82b00f614d6df56e6414118ba194a6a0f8413e5e