[test plan] Test plan for BGP scale test (#15702)
What is the motivation for this PR?
With numerous BGP sessions holding a lot routes, any flapping on BGP sessions or routes cloud have more overhead on device, to verify the functionality and estimate convergence time, we publish this test plan.

How did you do it?
Describe three test scenarios and introduce how we measure time in test.

Change-Id: I379b581c8341f09ff5a3d28342cd4e43c9ec5ed6