[bgp-scale-test] reorder test cases to avoid device unisolation test effect other tests (#18256)
What is the motivation for this PR?
The device unisolation test will restart all ports on DUT, after all bgp routes were stable in FRR after ports restart, however, the routes may not stable in ASIC, so the member scale test will be affected.

How did you do it?
To avoid this, move the device unisolation test to the last.

How did you verify/test it?
Run test on DUT

Change-Id: I653cc7c35e2abe68c3aec61a5480d137c2651f6d