fix LA marker injection at start of logs
Function insert_new_start_string() injects the LA test-start marker as
the very-first log-line in order for LA to run from the beginning of the
logs. It does so by artificially creating a new log file "older" than
all existing log files (by giving it a higher index-number than all
existing ones). The function also zips the file, but this is a bug: if
the file is syslog.1 then it shouldn't be zipped (only syslog.2 and
above are zipped). The LA was handling this ok, but during log rotation
this file gets overridden and deleted, which causes LA to fail if log
rotation occurred before the test ended.

Change-Id: Ic45dea116dceb7d685dc2740a09b0604b79740f2