Enhance PR test coverage for common functions in `tests/scripts`. (#18146)
What is the motivation for this PR?
There are no individual test scripts directly under the tests/scripts folder. However, this folder primarily contains common functions shared across various test cases. Previously, when changes were made within this folder, PR testing would be skipped because no standalone test scripts were detected.
In this PR, we improve the logic: if changes occur within the tests/scripts folder, we will now trigger the execution of the full PR test scope, as modifications to these common functions could potentially impact multiple test cases.

How did you do it?
We add a judgement that if the modified files are under the tests/scripts folder, we will trigger the execution of the full PR test scope.

How did you verify/test it?
Modify the scripts under the folder tests/scripts and run the PR test, we can see that the scope of PR test is the whole scope.

Change-Id: Ic00a3673706975f8c1ec3e2a325349f248822625