test_upgrade_path: Fix ceos bgp cmd parse (#18212)
What is the motivation for this PR?
Newer version of eos updates output of show ip bgp neighbor, which changes indentation. This breaks the parsing of restart_time.

How did you do it?
Regex match on integer and ignore indentation.

How did you verify/test it?
Tested on Arista testbed

Change-Id: I900982f6b7a70d3d6148cc4025b9ac79c2f2b3b5