Use is_alive over isAlive for py3 threads (#18213)
What is the motivation for this PR?
The python isAlive method for checking thread liveliness is for py2/earlier versions of py3. In py3.9 it is fully removed, and only is_alive is supported. Best to switch to is_alive for all versions of py3.

How did you do it?
How did you verify/test it?
Tested on Arista testbed

Change-Id: Id2ac58b89dff49dfdb2e575f82d1a41e5a862540