[tests_mark_conditions.yaml]: Skip GNMI/ZMQ test for isolated topology (#17985)
What is the motivation for this PR?
GNMI ZMQ has an issue in isolated topology, but fixing it is not a priority because it isn't used in isolated topology. So, just skip this test.

How did you do it?
Add a condition in the skip file

How did you verify/test it?
Check locally

Change-Id: I076f785ea307a601c27868c59edb11d53e6567c3