BGP testcases failure having sonic as neighbor device (#18300)
Fixes https://github.com/sonic-net/sonic-buildimage/issues/22578

The following are a set of testcase which fail when we use sonic
as neighbor device.  This set of testcases use EOS as neighbor
device.

List of test cases:
/tests/bgp/test_bgp_peer_shutdown.py
/tests/bgp/test_bgp_route_neigh_learning.py
/tests/bgp/test_bgp_bbr.py

The changes include some of the common and helper function
update to support sonic related configuration and validaton
and using these in these test cases.

Change-Id: Ifbccf858081f287238c69954647a277c1dd09b2c
Signed-off-by: Selvamani Ramasamy <selva@nexthop.ai>