Remove the skip/xfail for the dualtor_io link failure test (#18276)
Issue #16161 is fixed.
And some test cases that were failing in active-standby due to issue #8272 have been passing in active-active.
Only test_active_link_down_downstream_active_soc[active-active] is still failing, but the behavior is not exactly the same as active-standby. Probably we need to debug the failure again in active-active scenario.

Change-Id: Ib1c32599d389558bf8dd1f35f5e61007400c9fbc