Sonic mgmt tests for Per Vnet BGP and Dynamic peer modification (#17935)
What is the motivation for this PR?
The feature to allow BGP peers inside a vnet and being able to modify ip ranges for dynamic peers through bgpcfgd requires new set of testcases. This change addresses it.

Have a plan to enhance these sonic management tests in phase 2 with more testing on cross vnet communication, vxlan packet test. Will also test the impact of making changes in 1 vnet on other vnets.

How did you do it?
For per vnet BGP testcases:

Added a new template file which will be rendered to create the required VNET and BGP configurations on the DUT.
For modification of ip ranges, added 2 new templates to delete,add and modify ip ranges.
Validated using show bgp summary that the peer state is good and routes learnt are as expected.
How did you verify/test it?
Verified it on a sonic device by deploying the configuration and running the testcases.

Change-Id: Ia63849e27328ffb4b5615fca845fdf157a5a75f3