Updating telemetry metrics interfaces for metrics collection (#17973)
What is the motivation for this PR?
The current telemetry metrics interfaces is lacking the common metrics label and names for data collection, such as port/queue/psu/sensor and etc.

How did you do it?
This PR adds the most frequently used metrics labels as well as updating the examples for usage.

How did you verify/test it?
Validated with local influx db deployment.

Change-Id: Id07b90913303e472f965e5f5c9e29a30712603b6