Update setup_ntp_context to detect NTP daemon type (#18583)
Bookworm uses ntpsec as the NTP daemon on the PTF host. The PR is modified to be backward compatible with Bullseye. It detects the NTP daemon type which is NTP on Bullseye and NTPSEC on Bookworm. Fixes test failures on Bookworm that rely on the NTP daemon setup like ip/test_mgmt_ipv6_only.py::test_ntp_ipv6_only

Change-Id: Ia38c617c012346741f29a24b513b93fe5699f288