Modify gnmi stub generate script inline with protoc issue fixed (#18606)
Modify the generate gNMI stub script to check if the directory exists. This was due to a bug grpc/grpc#39583 in grpc_tools.protoc which is now fixed.

What is the motivation for this PR?
Fix issue with stub generation script.

How did you do it?
See description above.

How did you verify/test it?
Manually

Change-Id: I2bf158ceec11e78e62409a7112542ab2f8c44a4d