Adjust timeout and branch for internal pipeline context. (#18629)
What is the motivation for this PR?
The default timeout for both jobs and tasks is 60 minutes. However, collecting historical test cases can take longer than an hour, so we explicitly set a custom timeout for both the job and the task. At the same time, since this YAML file is synchronized from the public to the internal repository, but the pipeline actually runs in the internal repository, we need to update the branch name from 'master' to 'internal' to reflect the correct context.

How did you do it?
Add the timeout explicitly in each job and task and change the branch name from master to internal.

How did you verify/test it?
For test, we can refer to https://dev.azure.com/mssonic/internal/_build/results?buildId=860685&view=results

Change-Id: I55db42cdc872c7dc4967778d7e6698197d864f8f