Fix test_ntp_long_jump issue
The test initiates an NTP server on PTF, which has a pool of 4 network NTP servers and sets the local clock as the preferred option. However, the local server's clock (crystal oscillator) has poor accuracy, causing the PTF's NTP server to dynamically switch to other network NTP servers in the pool. As a result, when DUT uses the PTF NTP server as the clock, synchronization becomes difficult. To  resolve the issue, remove the network NTP configuration from the pool.

Change-Id: I3c3cb56005c50ed44660ce3511c38d31f4232a73