[dualtor-io] Fix duplication merge condition (#18828)
What is the motivation for this PR?
Work Item: 33161844

Actually, there is only one disruption, but due to the vlan flooding behavior, packet with id 70 is flooded in the vlan and received; so two disruptions are detected.
So for this case, the following two disruptions should be merged into one.

How did you do it?
Change the merge condition to allow merge in this case.

How did you verify/test it?
Run on Cisco dualtor testbed and pass.

Change-Id: I3c7fb55e734d7078e6fc9a3354a3e7fc83a86606
Signed-off-by: Longxiang Lyu <lolv@microsoft.com>