Valgrind first commit
Added methods for starting and stopping valgrind on a pre-defined list
of services, and collect valgrind output. Note this methods will replace
Tal's older code that's found in the same file, but meanwhile the older
code is not deleted until the new code will contain the full
functionality (running on dockers and on all services, and automatic
analysis or aggregation of the valgrind output).

Change-Id: I013fc3947bb255f387767e7d77cc3111efc20ae2