Retry SSH connection if fails after reboot, with logging for easier debugging. (#17816)
What is the motivation for this PR?
To improve debugging by adding logs and retries for SSH connection failures after a reboot.

How did you do it?
Adds a retry with a quick timeout and no expected "search_regex".
Adds logs to capture ping status.
Raises an exception if the first connection attempt fails.

Change-Id: I798b50952dc30d030e931fc7d5e73874c0a43cb7