fix PT0 template in t0-isolated topology (#18995)
Summary:
Two issues:
PT0 should use different template with T1 neighbors.
the current bp_interface subnet is: 10.10.246.0/22 (10.10.244.0-10.10.247.255). while in t0-isolated topology, the bp_interface may choose 10.10.248.2 or 10.10.248.3 which is outside the range.

What is the motivation for this PR?
Fix bp_interface and PT0 template for t0-isolated topology

Change-Id: I4555b033d635aefcc8e3881b447c2d9897927036