chore: fix link oper down (#18975)
Description of PR
Summary: Due to the new API changes, we're not using the variable link_state anymore. During recovery, we checked for this variable to perform recovery, however this is never set.

Fixes # (issue)

Approach
What is the motivation for this PR?
Fix the link oper down on IXIA test case

Change-Id: I9c35b214d052971add8d40048679b3f5a650a999
Signed-off-by: Austin Pham <austinpham@microsoft.com>