Remove special characters (special quotes) from Python scripts (#19040)
* Remove special characters (special quotes) from Python scripts

Remove non-standard quote marks from two Python scripts in the ansible
directory. For Python 3, things will still work. However, in cases where
these scripts are executed in a Python 2 context (because the SONiC
device is running an old version of SONiC), it will cause problems. This
may be the case for upgrade path testing, where old versions of SONiC
might be running on the device.

Change-Id: I28a7d22c62e1087a595f60a5b06f396d4bf87584
Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>