Collect techsupport on neighbors if fixture fails in pc/test_retry_count (#19035)
If the `disable_retry_count_on_peer` or `disable_retry_count_on_dut`
fixtures fail, then collect the techsupport on the neighbor devices.
There is already code to collect the techsupport on neighbor devices if
the main test fails, but not if the fixtures fail. However, the current
common cause of failures appear to be in the fixtures failing, which
means the current code won't catch that.

Change-Id: I441afa024d6b2b265db8fca41625099ec03cf105
Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>