[CI] Fix elastictest template by adding quotes to add-topo-params (#19085)
What is the motivation for this PR?
Previously, when setting ADD_TOPO_PARAMS, users had to wrap the value in double quotes inside single quotes, like this:

  - name: ADD_TOPO_PARAMS
    type: string
    default: "'-e enable_macsec=True -e macsec_profile=MACSEC_PROFILE'"
    displayName: "Add topo parameters"
This quoting pattern was error-prone and easy to get wrong by missing a quote.

How did you do it?
This PR moves the quoting into the template itself. With this change, users can simply provide the value as-is without worrying about extra quotes:

  - name: ADD_TOPO_PARAMS
    type: string
    default: "-e enable_macsec=True -e macsec_profile=MACSEC_PROFILE"
    displayName: "Add topo parameters"
How did you verify/test it?
I triggered a pipeline run and confirmed that the YAML parsed successfully and the parameters were applied correctly.

Change-Id: I1ade59f83a2db85b5969ad353efa4c963337a029