Updated the WRED ECN config update test so that it no longer requires a WRED profile named 'AZURE_LOSSLESS'. (#19246)
What is the motivation for this PR?
The test_ecn_config_update.py test fails on devices that do not have a WRED_PROFILE named AZURE_LOSSLESS.

How did you do it?
Instead of updating the WRED_PROFILE named AZURE_LOSSLESS, the test now updates all WRED profiles found in CONFIG DB and then verifies that these updates are applied to ASIC DB.
Note: In order for this test to pass, changes on the GCU side are also needed. Here is the PR in sonic-utilities for GCU changes: sonic-net/sonic-utilities#3910

How did you verify/test it?
Tested on a Mellanox switch with 3 WRED profiles, none of which were named AZURE_LOSSLESS. The old version of the test failed, while the new version passed.

Change-Id: I93a7b774cc86646e89e2b5e98f239200073da2fb
Signed-off-by: Mahdi Ramezani <mramezani@microsoft.com>