[GCU] Enhance test_eth_interface (#19341)
What is the motivation for this PR?
When interface FEC is not configured (N/A or none). The GCU rollback fails and port become operational down due to incorrect FEC.
Enhance the testcase to set FEC back to N/A (none).

How did you do it?
Enhance the testcase to set FEC back to N/A (none).

How did you verify/test it?
Verified by run test on Arista-7050CX3 M1 testbed.

Change-Id: Ifdd2101a85cca8bc9f8f7057c1d7472ad7278453