Add MD5 checksum verification for image deployment in Sonic CLI
Enhance the deploy_sonic method to include MD5 checksum validation for the image being copied to the device under test (DUT). This ensures the integrity of the image by comparing the local file's MD5 checksum with that of the copied file on the DUT, retrying the copy operation if they do not match.

Change-Id: Ie4b892cfbde66e181fa24d819a3baa530d85409e