SRv6: Update SRv6 test case to align with 16-bit locator function length (#19545)
What is the motivation for this PR?
Currently FRR limits the function length up to 20 bits for a srv6 locator, but the configuration in 7nodes_cisco topology is 32 bits. Let's change it as 16 (less than 20).

How did you do it?
Update the configuration for locator and SID in 7nodes_cisco, and update the test case for srv6 according to the new value.

How did you verify/test it?
This test case is a part of project Phoenix Wing, and the changes have been verified on its vSONiC testbed.

Change-Id: Ie645110aa98cf901e371104f99fd83eaaf0b69f1
Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>