Revert "Add new test gnmi 2038"
This reverts commit 9917487ef382f12f7bce1eb77e0fab0f363127ea.
This change is not merged in community for too long time and causing
conflicts from time to time. The latest conflict is with
https://github.com/sonic-net/sonic-mgmt/pull/19571, and it's impossible
to merge the changes without testing. Revert this change in develop for
now, if the test is still needed, will merge it back after the
community PR is merged.

Change-Id: I9975e80070612b2c04f623e89693b3cc3490953a