Use monit validate to get live data (#19759)
What is the motivation for this PR?
disk/test_disk_exhaustion.py creates a 1.7G file in the test and deletes it at the end of the test.
But "monit status" is configured to check only once every 60 secs in /etc/monit/monitrc.
This provides a stale data resulting in memory high threshold getting breached.

How did you do it?
We should use "monit validate" instead of "monit status"

How did you verify/test it?
verified by running the test

Change-Id: If6e2743c160bf2dcf1d146ad5844f44a7af3d873