[sonic-mgmt]Fix flakiness in pfcwd/test_pfcwd_all_port_storm.py for (#19705)
What is the motivation for this PR?
The test will pick a ToR randomly for the test without ascertaining if it's an active ToR
In such cases, when the selected ToR is the standby, resolve_arp(which sends out a docker exec swss arping <>) would fail since the downstream facing interface is in standby mode, and the test will not proceed
How did you do it?
This fix ensures the selected ToR will be set to be the active ToR

How did you verify/test it?
Verified on Arista-7050CX3 running dualtor topology

Change-Id: Iafb23be10cb2d7d079cd91accac595239c8cb6c1