[202505][sonic-mgmt]Fix for gnmi/test_mimic_hwproxy_cert_rotation.py (#19702)
What is the motivation for this PR?
The expression had an error and did not correctly match the feature status of telemetry/gnmi, leading it to incorrectly infer it as "Failed: Internal image has neither gnmi nor telemetry feature enabled"

How did you do it?
Fixing the expression to match for gnmi or telemetry under show feature status correctly

How did you verify/test it?
Tested it on Arista-7050CX3 devices and verified that it works

Any platform specific information?
Not platform specific

Change-Id: I70b1b97ddbce12e9dd9e46edc1b6c475c74cbb5d