Update restapi test to invoke apply_cert_config in smartswitch and isolated topology
The configuration needed for the restapi server is added directly to the running config using sonic-db-cli commands:
https://git-nbu-sw.nvidia.com/r/gitweb?p=switchx/sonic/sonic-mgmt.git;a=blob;f=tests/restapi/test_restapi.py;h=5be4008d7d4d353e69ff8f032faf2b0badf15a75;hb=refs/heads/develop#l52
And we perform warm-reboot:
https://git-nbu-sw.nvidia.com/r/gitweb?p=switchx/sonic/sonic-mgmt.git;a=blob;f=tests/restapi/test_restapi.py;h=5be4008d7d4d353e69ff8f032faf2b0badf15a75;hb=refs/heads/develop#l75
Which saves the running configuration to the config_db.json as part of the warmboot-finalizer.service script finalize-warmboot.sh
https://github.com/nvidia-sonic/sonic-buildimage/blob/0bd8460a403b77ba399a2f4f48e8aaff6c317932/files/image_config/warmboot-finalizer/finalize-warmboot.sh#L208
This causes all subsequent tests to pass as the restapi configuration is saved. In case of smartswitch the running configuration is not saved since warm-reboot is skipped.
So add a fixture to invoke apply_cert_config for smartswitch and
isolated topology

Change-Id: I37b90e3f455cc32e24adb1af73d90ab54e09d45c