[conditional mark] xfail test_generic_hash (#19955)
What is the motivation for this PR?
test_generic_hash is a test case that fails on many platforms and is only expected to be supported on a limited set. Since this feature is used in a very narrow scope, we’ll mark the test as xfail—it will still run, but the results won’t be validated for now.

How did you do it?
Add xfail for test_generic_test

How did you verify/test it?
N/A

Any platform specific information?
N/A

Change-Id: I1eedea026d39a7cca5718c03bdc02630332a9f64