bmp: Fix docker restart test for multi-asic (#19797)
Description of PR
Summary:
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Approach
What is the motivation for this PR?
There is a bmp container per ASIC so we need to use the correct per-ASIC names if we are on a multi-ASIC DUT.

How did you do it?
I updated the restart command to use the correct API for restarting a per-ASIC service.

How did you verify/test it?
We ran the test locally on an Arista multi-ASIC DUT.

Change-Id: I3baed9cc026e948ff566e87e01a1406209a5018f
signed-off-by: jianquanye@microsoft.com