Update ignore_regex in test_duplicate_route to support bulk update mode (#20357)
The sonic-sairedis change from sonic-net/sonic-sairedis#1613 modified the behavior of BULK operations which resulted in the log spew looking different before and after on API failures. This impacts test_duplicate_route.py. As this test is failing with the above sonic-sairedis change included, sonic-buildimage PRs that bump up the submodule reference fail pipeline checks.

The fix is to move all regexes to the ignore list so that the test can continue to pass with and without this sonic-sairedis change. Once the sairedis change merges, we can change the ignore list to the expect list.

Change-Id: Ia9bd6ca2f037833a9dd263fe1bfe1fc278dfc120