T2:snappi_tests:Change the speed_tol to correct number in pfcwd_actions. (#20074)
Description of PR
Summary:
The pfcwd_actions test: test_pfcwd_disable_pause_cngtn is being flaky due to setting the loss tolerance low. This is a regression introduced by #19772.

In this current PR, we are reverting the changes in traffic rate introduced by 19772 for this particular testcase.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Change-Id: I8f3784ea5c96d05ca6babd0523b70676930df131
signed-off-by: jianquanye@microsoft.com