Add dry run to DWRR weight change test (#20895)
Description of PR
This fixes testQosSaiDwrrWeightChange on Arista-7060X6-16PE-384C-B-O128S2. It was found that the test that typically runs previous to this one testQosSaiQSharedWatermark, sends a large amount of traffic to a single queue, which adds a bias to the scheduler and affects the results of this test. A similar dry run is already in place for testQosSaiDwrr.

Change-Id: I052a06b7045ee89c06db55e629511e33993d5888
signed-off-by: jianquanye@microsoft.com