Remove test_neighbour_mac from vpp test list (#20937)
Approach
What is the motivation for this PR?
test_neighbor_mac.py failed due to a race condition in setting IP on Ethernet0. This is because in response to removing Ethernet0 from PortChannel, sonic-vpp needs to update all the affected routes which takes a long time. In the end, setting IP on Ethernet0 happens before the interface is removed from the PortChannel. At the end, the IP on Ethernet0 was not set successfully so ping failed.
To fix this issue, sonic-vpp has to be more efficient in handling nexthop-group replacement. Until then, this test has to be disabled.

Change-Id: I446de89349410ce0c915352ba00076c8e122945b
signed-off-by: jianquanye@microsoft.com