Add PSU sensor test to validate sensor key prefix
Introduced a new test case in test_psu.py to ensure that the output from the PSU sensor has the expected prefix "PSU-". The test checks each sensor key, skipping the "Adapter" key and handling a specific case for the "temp3" sensor on the SN2700 platform. This enhancement improves the validation of sensor outputs in the system.

Change-Id: I35e5a25cff86aa808ed896ec55fb6d13c917d9bf