update ansible - develop
Update to devts tag 1.2.475, which have:
ansible-core 2.13.13
pytest-ansible 4.0.0

Also:
Sometime, we have the failure on canonical(SimX) setups, but not on
community:
AttributeError: module 'ansible.plugins.action.shell' has no attribute 'ActionModule'

It's happened as the versions not fully compatible and we using different
order of plugins loading.
The plugin loader in 2.13 initializes lazily - meaning the ActionModule class is not fully registered at the moment pytest first touches it.

Our WA:
manually define the class if it has not been loaded yet - effective when we can't upgrade Ansible

Change-Id: Ic1e9e59d25dddcaf8de19a42fd7e241a91df7ac3