Cisco-8000:Snappi:Fix the arp deletion code in cisco-8000 fixture. (#21064)
Description of PR
Summary:
In the cisco-specific function: gen_data_flow_dest_ip(), we also have a code to delete arp entries created by the same function. But that code is not really correct. It was not caught due to the try-except block protecting it. Essentially, the arguments to the arp delete CLI is wrong. This PR fixes this issue, with the correct arguments to the arp delete.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
 202405
 202411
 202505
Approach
What is the motivation for this PR?
The arp delete will not work with the current code. This PR is fixing the arguments to the arp delete.

How did you do it?
By calculating the correct arguments to the arp delete.

How did you verify/test it?
Ran it on T1 and T2 devices.

Any platform specific information?
Specific to cisco-8000 only.

@sdszhang , @auspham : FYI.

ssigned-off-by: jianquanye@microsoft.com
Change-Id: I225594e602cb82e22c4b0ffdce7d029001b751d0