Update sonic-mgmt tests for Trixie (#20839)
What is the motivation for this PR?
Update sonic-mgmt test cases for compatibility with Debian Trixie (and the changes that come with that), while maintaining compatibility with the current master branch image.

How did you do it?
How did you verify/test it?
KVM T0 and T1-lag tests on Trixie-based image (there are still some failures, but those require image changes):

https://elastictest.org/scheduler/publictestplan/690b90d58ad8ec69c27e475d
https://elastictest.org/scheduler/publictestplan/690b9101c30faf0722d8ae74

Change-Id: I50d4c2629e50b8ec185a569e95edae373517f0be
Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>