[xfail] a couple newly added test in multi-asic PR test are failing (#21692)
What is the motivation for this PR?
A couple test was added to multi-asic PR test suite and they are failing.

How did you do it?
This is not a fix. Since these are not recent tests, mark them xfail until they got addressed.

How did you verify/test it?
This PR test.

Change-Id: I2a7ddafc16c7cac844f740aa710265663827e347
Signed-off-by: Ying Xie <ying.xie@microsoft.com>