Update VsSetup documentation for folder permissions verifyation (#21766)
What is the motivation for this PR?
Follow the current document. I encountered permission error when adding topo t0.
image
The reason is that the home folder has a too strict permission code.

Add some note in the beginning of downloading cEOS image'. To prevent downloading the mismatch version.
How did you do it?
Add permission verification to VsSetup doc.

Change-Id: I2a13e5867d6aee0e85c01483700d6a108abb7816
Signed-off-by: cliffchen <t-cliffchen+github@microsoft.com>