Limiting the test to only VS, cisco and nvidia ASICs (#21737)
What is the motivation for this PR?
If this test is run on devices which do not support the feature, it can cause failures.

How did you do it?
Added a limitation to what platforms/ASICs this test should run on.

How did you verify/test it?
Ran on other asics to verify the test does not run. also ran on asics which its supported on to validate that it passes.

Change-Id: I1eeb35ca73288f45a0c589a0b49c1faa84b736a9